+++ SharePointDsc/DSCResources/MSFT_SPManagedMetadataServiceAppDefault/MSFT_SPManagedMetadataServiceAppDefault.psm1
@@ -10,9 +10,8 @@ function Get-TargetResource
     param
     (
         [Parameter(Mandatory = $true)]
-        [ValidateSet('Yes')]
         [String]
-        $IsSingleInstance,
+        $ServiceAppProxyGroup,
 
         [Parameter(Mandatory = $true)]
         [System.String]
@@ -35,11 +34,26 @@ function Get-TargetResource
 
         $params = $args[0]
 
-        $serviceAppProxies = Get-SPServiceApplicationProxy -ErrorAction SilentlyContinue
+        if ($params.ServiceAppProxyGroup -eq 'default')
+        {
+            $serviceAppProxyGroup = Get-SPServiceApplicationProxyGroup -Default -ErrorAction SilentlyContinue
+        }
+        else
+        {
+            $serviceAppProxyGroup = Get-SPServiceApplicationProxyGroup -Identity $params.ServiceAppProxyGroup `
+                -ErrorAction SilentlyContinue
+        }
+
+        if ($null -eq $serviceAppProxyGroup)
+        {
+            throw "Specified ServiceAppProxyGroup $($params.ServiceAppProxyGroup) does not exist."
+        }
+
+        $serviceAppProxies = $serviceAppProxyGroup.Proxies
 
         if ($null -eq $serviceAppProxies)
         {
-            throw "There are no Managed Metadata Service Application Proxy available in the farm"
+            throw "There are no Service Application Proxies available in the proxy group"
         }
 
         $serviceAppProxies = $serviceAppProxies | Where-Object -FilterScript {
@@ -48,7 +62,7 @@ function Get-TargetResource
 
         if ($null -eq $serviceAppProxies)
         {
-            throw "There are no Managed Metadata Service Application Proxy available in the farm"
+            throw "There are no Managed Metadata Service Application Proxies available in the proxy group"
         }
 
         $defaultSiteCollectionProxyIsSet = $false
@@ -86,10 +100,9 @@ function Get-TargetResource
         }
 
         return @{
-            IsSingleInstance               = $params.IsSingleInstance
+            ServiceAppProxyGroup           = $params.ServiceAppProxyGroup
             DefaultSiteCollectionProxyName = $defaultSiteCollectionProxy
             DefaultKeywordProxyName        = $defaultKeywordProxy
-            InstallAccount                 = $params.InstallAccount
         }
     }
     return $result
@@ -101,9 +114,8 @@ function Set-TargetResource
     param
     (
         [Parameter(Mandatory = $true)]
-        [ValidateSet('Yes')]
         [String]
-        $IsSingleInstance,
+        $ServiceAppProxyGroup,
 
         [Parameter(Mandatory = $true)]
         [System.String]
@@ -128,7 +140,17 @@ function Set-TargetResource
 
         $params = $args[0]
 
-        $serviceAppProxies = Get-SPServiceApplicationProxy -ErrorAction SilentlyContinue
+        if ($params.ServiceAppProxyGroup -eq 'default')
+        {
+            $serviceAppProxyGroup = Get-SPServiceApplicationProxyGroup -Default -ErrorAction SilentlyContinue
+        }
+        else
+        {
+            $serviceAppProxyGroup = Get-SPServiceApplicationProxyGroup -Identity $params.ServiceAppProxyGroup `
+                -ErrorAction SilentlyContinue
+        }
+
+        $serviceAppProxies = $serviceAppProxyGroup.Proxies
 
         $serviceAppProxies = $serviceAppProxies | Where-Object -FilterScript {
             $_.GetType().FullName -eq "Microsoft.SharePoint.Taxonomy.MetadataWebServiceApplicationProxy"
@@ -163,9 +185,8 @@ function Test-TargetResource
     param
     (
         [Parameter(Mandatory = $true)]
-        [ValidateSet('Yes')]
         [String]
-        $IsSingleInstance,
+        $ServiceAppProxyGroup,
 
         [Parameter(Mandatory = $true)]
         [System.String]